+++ app/controllers/admin/dashboard_controller.rb
@@ -1,8 +1,7 @@
 # frozen_string_literal: true
 
 module Admin
-  class DashboardController < ModuleController
-    def index
-    end
+  class DashboardController < BaseController
+    load_and_authorize_resource class: false
   end
 end
